pyproject.toml
[project]
name = "novel-knowledge-base"
version = "0.1.0"
description = "Novel Knowledge Base with MCP stdio server"
authors = [
    {name = "User", email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "volcengine-ark-runtime>=1.0.0",
]

[project.scripts]
novel-kb-mcp = "run_mcp_stdlib:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"